+++ tests/testthat/test-filesystem.R
@@ -22,3 +22,17 @@ test_that("using a filesystem cache works", {
   expect_true(is.memoised(fnm))
   expect_false(is.memoised(fn))
 })
+
+test_that("two functions with the same arguments produce different caches (#38)", {
+
+  temp <- tempfile()
+  fs <- cache_filesystem(temp)
+
+  f1 <- memoise(function() 1, cache = fs)
+  f2 <- memoise(function() 2, cache = fs)
+
+  expect_equal(f1(), 1)
+  expect_equal(f2(), 2)
+
+  expect_equal(length(list.files(temp)), 2)
+})